🐛 Fix: User로 통일하여 수정 및 mailsender 수정
