+++ src/main/java/dongguk/rangers/domain/user/UserRepository.java
@@ -1,12 +1,12 @@
 package dongguk.rangers.domain.user;
 
-import dongguk.rangers.domain.user.entity.Users;
+import dongguk.rangers.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Long> {
-    Optional<Users> findByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
 }
