+++ src/main/java/dongguk/rangers/domain/user/entity/User.java
@@ -2,12 +2,15 @@ package dongguk.rangers.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Entity
-@Table(name = "User")
+@Getter
+@Table(name = "users")
 @NoArgsConstructor
 @AllArgsConstructor
 public class User {
@@ -25,13 +28,25 @@ public class User {
     private String codeId;
 
     @Column(nullable = false)
-    private String name;
+    private String nickname;
 
     @Column(nullable = false)
     private String password;
 
     @Column(nullable = false)
     private String email;
+
+    // 카카오 로그인을 사용할 때 필요한 Builder
+    @Builder
+    public User(String email, String nickname) {
+        this.email = email;
+        this.nickname = nickname;
+    }
+
+    public void updateNickName(String nickname) {
+        this.nickname = nickname;
+    }
+
 
     // 보호자 역할일 때의 연결들
     @OneToMany(mappedBy = "guard")
@@ -40,4 +55,6 @@ public class User {
     // 피부양자 역할일 때의 연결들
     @OneToMany(mappedBy = "dependant")
     private List<Connect> dependantConnections;
+
+
 }
