+++ src/main/resources/application.properties
@@ -1,14 +1,14 @@
 # ??? ??
-kakao.client_id=${KAKAO_CLIENT_ID}
-kakao.client_secret=${KAKAO_CLIENT_SECRET}
+kakao.client_id=8febc48e07cfb2360b3973bf7aa1180a
+kakao.client_secret=QoJ6IemdOefC8NJmGjqP41tPxxM6TC7G
 kakao.redirect_uri=http://localhost:8080/api/kakao/callback
 kakao.get_code_path=https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=
 
 
 # MySQL ?????? ?? ??
-spring.datasource.url=${SPRING_DATASOURCE_URL}
-spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
-spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
+spring.datasource.url=jdbc:mysql://db-rangers.cdsys6gw2jyv.ap-northeast-2.rds.amazonaws.com:3306/cocofarm
+spring.datasource.username=rangers
+spring.datasource.password=farmsystem
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # JPA ??
@@ -18,8 +18,16 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.properties.hibernate.format_sql=true
 
 # Redis ?? ??
-spring.redis.host=${SPRING_REDIS_HOST}
+spring.redis.host=redis-keerbu.serverless.apn2.cache.amazonaws.com:6379
 spring.redis.port=6379
 
 # JWT ??
-jwt.secret=${JWT_SECRET}
+jwt.secret==qR5PFaKmHpcgCG/pEIllMG2tl41qAJYkmO6/vjz+3Jk=
+
+# mail ??
+spring.mail.host=${MAIL_HOST}
+spring.mail.port=587
+spring.mail.username=${MAIL_ID}
+spring.mail.password=${MAIL_PASSWORD}
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
